Modified the instantiation of camera amulating a polyfill
